Cast usleep parameter to Int
This request is linked to https://github.com/ccxt/ccxt/issues/1850

+++ php/Exchange.php
@@ -718,7 +718,7 @@ abstract class Exchange {
         $elapsed = $now - $this->lastRestRequestTimestamp;
         if ($elapsed < $this->rateLimit) {
             $delay = $this->rateLimit - $elapsed;
-            usleep ($delay * 1000.0);
+              usleep ((int)($delay * 1000.0));
         }
     }
